requirements.txt
# Core Flask
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
alembic>=1.13.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Visualization (for Streamlit compatibility)
streamlit>=1.30.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Redis (optional for caching)
redis>=5.0.0

# Celery (optional for background tasks)
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# Code Quality
flake8>=7.0.0
black>=24.0.0
mypy>=1.8.0

# Production Server
gunicorn>=21.2.0

# Security
werkzeug>=3.0.0
